Update clean command

+++ src/doc.rs
@@ -125,17 +125,16 @@ impl Doc {
         Ok(())
     }
 
-    pub fn clean_project(&self) -> Result<(), std::io::Error> {
-        // TODO
-        // if distclean {
-        // we can just remove output directory in the doc path
-        let mut output_dir = self.path.to_path_buf();
-        output_dir.push("output");
-        fs::remove_dir_all(&output_dir.as_path())?;
-        // } else {
-        // ... call make clean
-        //do_cmd("make", &["clean"])?;
-        // }
+    pub fn clean_project(&self, distclean: bool) -> Result<(), std::io::Error> {
+        if distclean {
+            // we can just remove output directory in the doc path
+            let mut output_dir = self.path.to_path_buf();
+            output_dir.push("output");
+            fs::remove_dir_all(&output_dir.as_path())?;
+        } else {
+            // ... call make clean
+            do_cmd("make", &["clean"])?;
+        }
 
         Ok(())
     }
